app/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.0.4
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
protobuf==3.15.8
six==1.15.0
SQLAlchemy==1.4.7
typing-extensions==3.7.4.3
uWSGI==2.0.18
Werkzeug==1.0.1
zipp==3.4.1
